Move dialog move to data/dialogs, routes to navigation/
